plotty: plot: refactor stdin/stdout support
Support for stdin/stdout was implemented before this change by using
pre-open file descriptors (`sys.stdin` and `sys.stdout`) when the input/output
was selected as "-".

After this change, we use instead '/dev/fd/0' for stdin and '/dev/fd/1'
for stdout, which can be open.

This fix some minor issues, and removes the last flake8 complaint.